aws - network attached resource ingress/egress filters
