+++ c7n/filters/__init__.py
@@ -22,5 +22,5 @@ from .health import HealthEventFilter
 from .iamaccess import CrossAccountAccessFilter, PolicyChecker
 from .iamanalyzer import AccessAnalyzer
 from .metrics import MetricsFilter, ShieldMetrics
-# from .vpc import DefaultVpcBase
+from .vpc import DefaultVpcBase
 from .waf import WafV2FilterBase, WafClassicRegionalFilterBase
