+++ c7n/resources/vpc.py
@@ -1248,423 +1248,8 @@ class SGDefaultVpc(net_filters.DefaultVpcBase):
         return self.match(resource['VpcId'])
 
 
-class SGPermission(Filter):
-    """Filter for verifying security group ingress and egress permissions
-
-    All attributes of a security group permission are available as
-    value filters.
-
-    If multiple attributes are specified the permission must satisfy
-    all of them. Note that within an attribute match against a list value
-    of a permission we default to or.
-
-    If a group has any permissions that match all conditions, then it
-    matches the filter.
-
-    Permissions that match on the group are annotated onto the group and
-    can subsequently be used by the remove-permission action.
-
-    We have specialized handling for matching `Ports` in ingress/egress
-    permission From/To range. The following example matches on ingress
-    rules which allow for a range that includes all of the given ports.
-
-    .. code-block:: yaml
-
-      - type: ingress
-        Ports: [22, 443, 80]
-
-    As well for verifying that a rule only allows for a specific set of ports
-    as in the following example. The delta between this and the previous
-    example is that if the permission allows for any ports not specified here,
-    then the rule will match. ie. OnlyPorts is a negative assertion match,
-    it matches when a permission includes ports outside of the specified set.
-
-    .. code-block:: yaml
-
-      - type: ingress
-        OnlyPorts: [22]
-
-    For simplifying ipranges handling which is specified as a list on a rule
-    we provide a `Cidr` key which can be used as a value type filter evaluated
-    against each of the rules. If any iprange cidr match then the permission
-    matches.
-
-    .. code-block:: yaml
-
-      - type: ingress
-        IpProtocol: -1
-        FromPort: 445
-
-    We also have specialized handling for matching self-references in
-    ingress/egress permissions. The following example matches on ingress
-    rules which allow traffic its own same security group.
-
-    .. code-block:: yaml
-
-      - type: ingress
-        SelfReference: True
-
-    As well for assertions that a ingress/egress permission only matches
-    a given set of ports, *note* OnlyPorts is an inverse match.
-
-    .. code-block:: yaml
-
-      - type: egress
-        OnlyPorts: [22, 443, 80]
-
-      - type: egress
-        Cidr:
-          value_type: cidr
-          op: in
-          value: x.y.z
-
-    `value_type: cidr` can also filter if cidr is a subset of cidr
-    value range. In this example we are allowing any smaller cidrs within
-    allowed_cidrs.csv.
-
-    .. code-block:: yaml
-
-      - type: ingress
-        Cidr:
-          value_type: cidr
-          op: not-in
-          value_from:
-            url: s3://a-policy-data-us-west-2/allowed_cidrs.csv
-            format: csv
-
-    or value can be specified as a list.
-
-    .. code-block:: yaml
-
-      - type: ingress
-        Cidr:
-          value_type: cidr
-          op: not-in
-          value: ["10.0.0.0/8", "192.168.0.0/16"]
-
-    `Cidr` can match ipv4 rules and `CidrV6` can match ipv6 rules.  In
-    this example we are blocking global inbound connections to SSH or
-    RDP.
-
-    .. code-block:: yaml
-
-      - or:
-        - type: ingress
-          Ports: [22, 3389]
-          Cidr:
-            value: "0.0.0.0/0"
-        - type: ingress
-          Ports: [22, 3389]
-          CidrV6:
-            value: "::/0"
-
-    `SGReferences` can be used to filter out SG references in rules.
-    In this example we want to block ingress rules that reference a SG
-    that is tagged with `Access: Public`.
-
-    .. code-block:: yaml
-
-      - type: ingress
-        SGReferences:
-          key: "tag:Access"
-          value: "Public"
-          op: equal
-
-    We can also filter SG references based on the VPC that they are
-    within. In this example we want to ensure that our outbound rules
-    that reference SGs are only referencing security groups within a
-    specified VPC.
-
-    .. code-block:: yaml
-
-      - type: egress
-        SGReferences:
-          key: 'VpcId'
-          value: 'vpc-11a1a1aa'
-          op: equal
-
-    Likewise, we can also filter SG references by their description.
-    For example, we can prevent egress rules from referencing any
-    SGs that have a description of "default - DO NOT USE".
-
-    .. code-block:: yaml
-
-      - type: egress
-        SGReferences:
-          key: 'Description'
-          value: 'default - DO NOT USE'
-          op: equal
-
-    By default, this filter matches a security group rule if
-    _all_ of its keys match. Using `match-operator: or` causes a match
-    if _any_ key matches. This can help consolidate some simple
-    cases that would otherwise require multiple filters. To find
-    security groups that allow all inbound traffic over IPv4 or IPv6,
-    for example, we can use two filters inside an `or` block:
-
-    .. code-block:: yaml
-
-      - or:
-        - type: ingress
-          Cidr: "0.0.0.0/0"
-        - type: ingress
-          CidrV6: "::/0"
-
-    or combine them into a single filter:
-
-    .. code-block:: yaml
-
-      - type: ingress
-        match-operator: or
-          Cidr: "0.0.0.0/0"
-          CidrV6: "::/0"
-
-    Note that evaluating _combinations_ of factors (e.g. traffic over
-    port 22 from 0.0.0.0/0) still requires separate filters.
-    """
-
-    perm_attrs = {
-        'IpProtocol', 'FromPort', 'ToPort', 'UserIdGroupPairs',
-        'IpRanges', 'PrefixListIds'}
-    filter_attrs = {
-        'Cidr', 'CidrV6', 'Ports', 'OnlyPorts',
-        'SelfReference', 'Description', 'SGReferences'}
-    attrs = perm_attrs.union(filter_attrs)
-    attrs.add('match-operator')
-    attrs.add('match-operator')
-
-    def validate(self):
-        delta = set(self.data.keys()).difference(self.attrs)
-        delta.remove('type')
-        if delta:
-            raise PolicyValidationError("Unknown keys %s on %s" % (
-                ", ".join(delta), self.manager.data))
-        return self
-
-    def process(self, resources, event=None):
-        self.vfilters = []
-        fattrs = list(sorted(self.perm_attrs.intersection(self.data.keys())))
-        self.ports = 'Ports' in self.data and self.data['Ports'] or ()
-        self.only_ports = (
-            'OnlyPorts' in self.data and self.data['OnlyPorts'] or ())
-        for f in fattrs:
-            fv = self.data.get(f)
-            if isinstance(fv, dict):
-                fv['key'] = f
-            else:
-                fv = {f: fv}
-            vf = ValueFilter(fv, self.manager)
-            vf.annotate = False
-            self.vfilters.append(vf)
-        return super(SGPermission, self).process(resources, event)
-
-    def process_ports(self, perm):
-        found = None
-        if 'FromPort' in perm and 'ToPort' in perm:
-            for port in self.ports:
-                if port >= perm['FromPort'] and port <= perm['ToPort']:
-                    found = True
-                    break
-                found = False
-            only_found = False
-            for port in self.only_ports:
-                if port == perm['FromPort'] and port == perm['ToPort']:
-                    only_found = True
-            if self.only_ports and not only_found:
-                found = found is None or found and True or False
-            if self.only_ports and only_found:
-                found = False
-        return found
-
-    def _process_cidr(self, cidr_key, cidr_type, range_type, perm):
-
-        found = None
-        ip_perms = perm.get(range_type, [])
-        if not ip_perms:
-            return False
-
-        match_range = self.data[cidr_key]
-
-        if isinstance(match_range, dict):
-            match_range['key'] = cidr_type
-        else:
-            match_range = {cidr_type: match_range}
-
-        vf = ValueFilter(match_range, self.manager)
-        vf.annotate = False
-
-        for ip_range in ip_perms:
-            found = vf(ip_range)
-            if found:
-                break
-            else:
-                found = False
-        return found
-
-    def process_cidrs(self, perm):
-        found_v6 = found_v4 = None
-        if 'CidrV6' in self.data:
-            found_v6 = self._process_cidr('CidrV6', 'CidrIpv6', 'Ipv6Ranges', perm)
-        if 'Cidr' in self.data:
-            found_v4 = self._process_cidr('Cidr', 'CidrIp', 'IpRanges', perm)
-        match_op = self.data.get('match-operator', 'and') == 'and' and all or any
-        cidr_match = [k for k in (found_v6, found_v4) if k is not None]
-        if not cidr_match:
-            return None
-        return match_op(cidr_match)
-
-    def process_description(self, perm):
-        if 'Description' not in self.data:
-            return None
-
-        d = dict(self.data['Description'])
-        d['key'] = 'Description'
-
-        vf = ValueFilter(d, self.manager)
-        vf.annotate = False
-
-        for k in ('Ipv6Ranges', 'IpRanges', 'UserIdGroupPairs', 'PrefixListIds'):
-            if k not in perm or not perm[k]:
-                continue
-            return vf(perm[k][0])
-        return False
-
-    def process_self_reference(self, perm, sg_id):
-        found = None
-        ref_match = self.data.get('SelfReference')
-        if ref_match is not None:
-            found = False
-        if 'UserIdGroupPairs' in perm and 'SelfReference' in self.data:
-            self_reference = sg_id in [p['GroupId']
-                                       for p in perm['UserIdGroupPairs']]
-            if ref_match is False and not self_reference:
-                found = True
-            if ref_match is True and self_reference:
-                found = True
-        return found
-
-    def process_sg_references(self, perm, owner_id):
-        sg_refs = self.data.get('SGReferences')
-        if not sg_refs:
-            return None
-
-        sg_perm = perm.get('UserIdGroupPairs', [])
-        if not sg_perm:
-            return False
-
-        sg_group_ids = [p['GroupId'] for p in sg_perm if p.get('UserId', '') == owner_id]
-        sg_resources = self.manager.get_resources(sg_group_ids)
-        vf = ValueFilter(sg_refs, self.manager)
-        vf.annotate = False
-
-        for sg in sg_resources:
-            if vf(sg):
-                return True
-        return False
-
-    def expand_permissions(self, permissions):
-        """Expand each list of cidr, prefix list, user id group pair
-        by port/protocol as an individual rule.
-
-        The console ux automatically expands them out as addition/removal is
-        per this expansion, the describe calls automatically group them.
-        """
-        for p in permissions:
-            np = dict(p)
-            values = {}
-            for k in (u'IpRanges',
-                      u'Ipv6Ranges',
-                      u'PrefixListIds',
-                      u'UserIdGroupPairs'):
-                values[k] = np.pop(k, ())
-                np[k] = []
-            for k, v in values.items():
-                if not v:
-                    continue
-                for e in v:
-                    ep = dict(np)
-                    ep[k] = [e]
-                    yield ep
-
-    def __call__(self, resource):
-        matched = []
-        sg_id = resource['GroupId']
-        owner_id = resource['OwnerId']
-        match_op = self.data.get('match-operator', 'and') == 'and' and all or any
-        for perm in self.expand_permissions(resource[self.ip_permissions_key]):
-            perm_matches = {}
-            for idx, f in enumerate(self.vfilters):
-                perm_matches[idx] = bool(f(perm))
-            perm_matches['description'] = self.process_description(perm)
-            perm_matches['ports'] = self.process_ports(perm)
-            perm_matches['cidrs'] = self.process_cidrs(perm)
-            perm_matches['self-refs'] = self.process_self_reference(perm, sg_id)
-            perm_matches['sg-refs'] = self.process_sg_references(perm, owner_id)
-            perm_match_values = list(filter(
-                lambda x: x is not None, perm_matches.values()))
-
-            # account for one python behavior any([]) == False, all([]) == True
-            if match_op == all and not perm_match_values:
-                continue
-
-            match = match_op(perm_match_values)
-            if match:
-                matched.append(perm)
-
-        if matched:
-            resource.setdefault('Matched%s' % self.ip_permissions_key, []).extend(matched)
-            return True
-
-
-SGPermissionSchema = {
-    'match-operator': {'type': 'string', 'enum': ['or', 'and']},
-    'Ports': {'type': 'array', 'items': {'type': 'integer'}},
-    'SelfReference': {'type': 'boolean'},
-    'OnlyPorts': {'type': 'array', 'items': {'type': 'integer'}},
-    'IpProtocol': {
-        'oneOf': [
-            {'enum': ["-1", -1, 'tcp', 'udp', 'icmp', 'icmpv6']},
-            {'$ref': '#/definitions/filters/value'}
-        ]
-    },
-    'FromPort': {'oneOf': [
-        {'$ref': '#/definitions/filters/value'},
-        {'type': 'integer'}]},
-    'ToPort': {'oneOf': [
-        {'$ref': '#/definitions/filters/value'},
-        {'type': 'integer'}]},
-    'UserIdGroupPairs': {},
-    'IpRanges': {},
-    'PrefixListIds': {},
-    'Description': {},
-    'Cidr': {},
-    'CidrV6': {},
-    'SGReferences': {}
-}
-
-
-@SecurityGroup.filter_registry.register('ingress')
-class IPPermission(SGPermission):
-
-    ip_permissions_key = "IpPermissions"
-    schema = {
-        'type': 'object',
-        'additionalProperties': False,
-        'properties': {'type': {'enum': ['ingress']}},
-        'required': ['type']}
-    schema['properties'].update(SGPermissionSchema)
-
-
-@SecurityGroup.filter_registry.register('egress')
-class IPPermissionEgress(SGPermission):
-
-    ip_permissions_key = "IpPermissionsEgress"
-    schema = {
-        'type': 'object',
-        'additionalProperties': False,
-        'properties': {'type': {'enum': ['egress']}},
-        'required': ['type']}
-    schema['properties'].update(SGPermissionSchema)
+SecurityGroup.filter_registry.register('ingress', net_filters.IPPermission)
+SecurityGroup.filter_registry.register('egress', net_filters.IPPermissionEgress)
 
 
 @SecurityGroup.action_registry.register('delete')
